+++ src/main/java/dk/aau/cs/giraf/train/opengl/PictoFrameLayout.java
@@ -16,59 +16,59 @@ import dk.aau.cs.giraf.pictogram.Pictogram;
  */
 public class PictoFrameLayout extends FrameLayout {
 
-	public PictoFrameLayout(Context context){
-		super(context);
-	}
-	
-	public PictoFrameLayout(Context context, AttributeSet attrs) {
-		super(context, attrs);
-	}
-	
-	public void addPictogramsToFrames(int pictoId){
-		Pictogram pic = PictoFactory.getPictogram(super.getContext(), pictoId);
-		pic.setOnTouchListener(new TouchListener());
-		pic.renderAll();
-		
-		PictoFrameLayout.LayoutParams frameLayoutParams = new PictoFrameLayout.LayoutParams(
-				android.view.ViewGroup.LayoutParams.MATCH_PARENT,
-				android.view.ViewGroup.LayoutParams.MATCH_PARENT);
-		
-		this.addView(pic, frameLayoutParams);
-		this.setTag("filled");
-		this.invalidate();
-	}
-	
-	public Pictogram getPictogram(){
-		return (Pictogram)this.getChildAt(0);
-	}
-	
-	/**
-	 * A touch listener that starts a drag event. There should also be a
-	 * receiver implementing {@link OnDragListener}.
-	 * 
-	 * see DragListener
-	 */
-	private final class TouchListener implements OnTouchListener {
-		@Override
-		public boolean onTouch(View view, MotionEvent motionEvent) {
-			if (motionEvent.getAction() == MotionEvent.ACTION_DOWN) {
-				ClipData data = ClipData.newPlainText("", "");
-				DragShadowBuilder shadowBuilder = new View.DragShadowBuilder(view);
-				view.startDrag(data, shadowBuilder, view, 0);
-				view.setVisibility(View.INVISIBLE);
-				return true;
-			}
-			else if (motionEvent.getAction() == MotionEvent.ACTION_UP) {// prevents that a picto disapears if only pressed and no drag
-				if(view != null && view.getVisibility() == View.INVISIBLE){
-					view.setVisibility(View.VISIBLE);
-				}
-				return true;
-			}
-			else {
-				return false;
-			}
-		}
-		
-	}
+    public PictoFrameLayout(Context context){
+        super(context);
+    }
+
+    public PictoFrameLayout(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public void addPictogramsToFrames(long pictoId){
+        Pictogram pic = PictoFactory.getPictogram(super.getContext(), pictoId);
+        pic.setOnTouchListener(new TouchListener());
+        pic.renderAll();
+
+        PictoFrameLayout.LayoutParams frameLayoutParams = new PictoFrameLayout.LayoutParams(
+                android.view.ViewGroup.LayoutParams.MATCH_PARENT,
+                android.view.ViewGroup.LayoutParams.MATCH_PARENT);
+
+        this.addView(pic, frameLayoutParams);
+        this.setTag("filled");
+        this.invalidate();
+    }
+
+    public Pictogram getPictogram(){
+        return (Pictogram)this.getChildAt(0);
+    }
+
+    /**
+     * A touch listener that starts a drag event. There should also be a
+     * receiver implementing {@link OnDragListener}.
+     *
+     * see DragListener
+     */
+    private final class TouchListener implements OnTouchListener {
+        @Override
+        public boolean onTouch(View view, MotionEvent motionEvent) {
+            if (motionEvent.getAction() == MotionEvent.ACTION_DOWN) {
+                ClipData data = ClipData.newPlainText("", "");
+                DragShadowBuilder shadowBuilder = new View.DragShadowBuilder(view);
+                view.startDrag(data, shadowBuilder, view, 0);
+                view.setVisibility(View.INVISIBLE);
+                return true;
+            }
+            else if (motionEvent.getAction() == MotionEvent.ACTION_UP) {// prevents that a picto disapears if only pressed and no drag
+                if(view != null && view.getVisibility() == View.INVISIBLE){
+                    view.setVisibility(View.VISIBLE);
+                }
+                return true;
+            }
+            else {
+                return false;
+            }
+        }
+
+    }
 
 }
